Added states

+++ code/collisions.py
@@ -3,9 +3,12 @@ import pygame
 from australian import *
 from projectiles import *
 
-def collision_x(target1, target2, iterable):
+def collision_x(target1, target2, iterable, state):
     if iterable:
-        for item1 in target1:
+        for item1 in target1: 
+            if item1.state != state: # used for enemies, the player transcends states, so the else case doesn't have it
+                continue
+
             for item2 in target2:
                 if item1.rect.colliderect(item2.rect):
                     if item1.direction.x > 0:
@@ -20,9 +23,12 @@ def collision_x(target1, target2, iterable):
                 elif target1.direction.x < 0:
                     target1.rect.left = item.rect.right
 
-def collision_y(target1, target2, iterable):
+def collision_y(target1, target2, iterable, state):
     if iterable:
         for item1 in target1:
+            if item1.state != state:
+                continue
+            
             for item2 in target2:
                 if item1.rect.colliderect(item2.rect):
                     if item1.direction.y > 0:
@@ -38,8 +44,11 @@ def collision_y(target1, target2, iterable):
                     target1.rect.top = item.rect.bottom
 
 
-def collision_projectile(projectiles, enemies, walls):
+def collision_projectile(projectiles, enemies, walls, state):
     for projectile in projectiles:
+        if projectile.state != state:
+            continue
+
         for wall in walls:
             if projectile.rect.colliderect(wall):
                 projectile.kill()
@@ -49,9 +58,12 @@ def collision_projectile(projectiles, enemies, walls):
                 enemy.health -= 100
                 projectile.kill()
 
-def check_enemy_projectiles(player, powerups, powerup_timers, enemy_projectile_group, walls):
+def check_enemy_projectiles(player, powerups, powerup_timers, enemy_projectile_group, walls, state):
     now = pygame.time.get_ticks()
     for projectile in enemy_projectile_group:
+        if projectile.state != state:
+            continue
+
         if projectile.rect.colliderect(player.rect):
             if type(projectile) == Beer:
                 powerups["drunk"] = 0
@@ -62,9 +74,12 @@ def check_enemy_projectiles(player, powerups, powerup_timers, enemy_projectile_g
             if projectile.rect.colliderect(wall.rect):
                 projectile.kill()
 
-def le_attack(player, enemy_group, powerups, powerup_timers):
+def le_attack(player, enemy_group, powerups, powerup_timers, state):
     now = pygame.time.get_ticks()
     for enemy in enemy_group:
+        if enemy.state != state:
+            continue
+
         if enemy.can_attack_primary and enemy.rect.colliderect(player.rect):
             player.health -= enemy.attack
             enemy.can_attack_primary = False
